requirements.txt
opcua==0.98.12
influxdb==5.3.1
PyYAML==5.3.1
pytest==6.2.1
jsonschema==3.2.0